Add component for random greeting
