--- src/Greeting.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function Greeting() {
-  return (
-    <div>Greeting</div>
-  );
-}
-
-export default Greeting;
